requirements.txt
deepchem==2.6.1
keras==2.10.0
numpy==1.23.2
pandas==1.5.2
Pillow==9.3.0
rdkit==2022.9.1
scikit_learn==1.1.3
streamlit==1.15.1
tensorflow==2.11.0